changed input data model to float64, adds AddAction tests

+++ pkg/actions/models.go
@@ -4,7 +4,7 @@ package actions
 // serialized JSON string -> JSON struct conversion
 type Input struct {
 	Action string `json:"action"`
-	Time   int    `json:"time"`
+	Time   float64    `json:"time"`
 }
 
 // Data model for temp local data store.
